requirements.txt
pandas
numpy
matplotlib
scikit-learn
torch
torchvision
tqdm
pyyaml
